dim/requirements.txt
# Workaround for https://github.com/pypa/setuptools/issues/940
setuptools<34.0.0

mysqlclient==1.3.8
SQLAlchemy==1.1.1
pyldap==2.4.25.1
Werkzeug==0.11.11
Flask==0.10
Flask-SQLAlchemy==2.1
Jinja2==2.6
Flask-Script==0.5.3
dnspython==1.15.0
simplejson==2.6.1
argparse==1.2.1
requests==2.8.1
pycrypto==2.6.1
six==1.10.0

# Python2
ipaddress==1.0.17;python_version<'3.0'
